Improve Debugger Tests ART Integration
Extract an interface that will allow Android Studio to inject a class that will handle ART VM debugger dependencies.

The interface is agnostic to ART. It should work for any remote VM.
