+++ plugins/kotlin/jvm-debugger/test/test/org/jetbrains/kotlin/idea/debugger/test/JvmAttacher.kt
@@ -1,0 +1,44 @@
+// Copyright 2000-2024 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license.
+package org.jetbrains.kotlin.idea.debugger.test
+
+import com.intellij.debugger.impl.DebuggerSession
+import com.intellij.debugger.impl.GenericDebuggerRunnerSettings
+import com.intellij.debugger.impl.RemoteConnectionBuilder
+import com.intellij.execution.configurations.JavaCommandLineState
+import com.intellij.execution.configurations.JavaParameters
+import com.intellij.execution.runners.ExecutionEnvironment
+import com.intellij.execution.target.TargetEnvironmentRequest
+import com.intellij.execution.target.TargetedCommandLineBuilder
+
+/** Attaches to a Java VM */
+internal class JvmAttacher : VmAttacher {
+
+    override fun attachVirtualMachine(
+        testCase: KotlinDescriptorTestCase,
+        javaParameters: JavaParameters,
+        environment: ExecutionEnvironment
+    ): DebuggerSession {
+        val debuggerRunnerSettings = (environment.runnerSettings as GenericDebuggerRunnerSettings)
+        val javaCommandLineState: JavaCommandLineState = object : JavaCommandLineState(environment) {
+            override fun createJavaParameters() = javaParameters
+
+            override fun createTargetedCommandLine(request: TargetEnvironmentRequest): TargetedCommandLineBuilder {
+                return getJavaParameters().toCommandLine(request)
+            }
+        }
+
+        val debugParameters =
+            RemoteConnectionBuilder(
+                debuggerRunnerSettings.LOCAL,
+                debuggerRunnerSettings.transport,
+                debuggerRunnerSettings.debugPort
+            )
+                .checkValidity(true)
+                .asyncAgent(true)
+                .create(javaCommandLineState.javaParameters)
+
+        val env = javaCommandLineState.environment
+
+        return testCase.attachVirtualMachine(javaCommandLineState, env, debugParameters, false)
+    }
+}
