+++ plugins/kotlin/jvm-debugger/test/test/org/jetbrains/kotlin/idea/debugger/test/KotlinDescriptorTestCase.kt
@@ -5,25 +5,22 @@ package org.jetbrains.kotlin.idea.debugger.test
 import com.intellij.debugger.DebuggerManagerEx
 import com.intellij.debugger.DefaultDebugEnvironment
 import com.intellij.debugger.engine.DebugProcessImpl
-import com.intellij.debugger.engine.RemoteStateState
-import com.intellij.debugger.impl.*
+import com.intellij.debugger.impl.DebuggerSession
+import com.intellij.debugger.impl.DescriptorTestCase
+import com.intellij.debugger.impl.GenericDebuggerRunnerSettings
+import com.intellij.debugger.impl.OutputChecker
 import com.intellij.debugger.settings.DebuggerSettings
 import com.intellij.execution.ExecutionTestCase
-import com.intellij.execution.configurations.JavaCommandLineState
 import com.intellij.execution.configurations.JavaParameters
 import com.intellij.execution.executors.DefaultDebugExecutor
 import com.intellij.execution.process.ProcessAdapter
 import com.intellij.execution.process.ProcessEvent
 import com.intellij.execution.process.ProcessOutputTypes
-import com.intellij.execution.runners.ExecutionEnvironment
 import com.intellij.execution.runners.ExecutionEnvironmentBuilder
-import com.intellij.execution.target.TargetEnvironmentRequest
-import com.intellij.execution.target.TargetedCommandLineBuilder
 import com.intellij.openapi.application.invokeAndWaitIfNeeded
 import com.intellij.openapi.application.runInEdt
 import com.intellij.openapi.application.runReadAction
 import com.intellij.openapi.application.runWriteAction
-import com.intellij.openapi.observable.util.whenDisposed
 import com.intellij.openapi.roots.LibraryOrderEntry
 import com.intellij.openapi.roots.ModifiableRootModel
 import com.intellij.openapi.roots.ModuleRootManager
@@ -37,7 +34,6 @@ import com.intellij.testFramework.IndexingTestUtil
 import com.intellij.testFramework.runInEdtAndGet
 import com.intellij.util.ThrowableRunnable
 import com.intellij.util.containers.addIfNotNull
-import com.intellij.util.io.delete
 import com.intellij.xdebugger.XDebugSession
 import org.jetbrains.kotlin.config.*
 import org.jetbrains.kotlin.fileClasses.JvmFileClassUtil
@@ -51,10 +47,6 @@ import org.jetbrains.kotlin.idea.compiler.configuration.KotlinCommonCompilerArgu
 import org.jetbrains.kotlin.idea.compiler.configuration.KotlinCompilerSettings
 import org.jetbrains.kotlin.idea.compiler.configuration.KotlinPluginLayout
 import org.jetbrains.kotlin.idea.debugger.evaluate.KotlinEvaluator
-import org.jetbrains.kotlin.idea.debugger.test.ArtUtils.buildCommandLine
-import org.jetbrains.kotlin.idea.debugger.test.ArtUtils.buildDexFile
-import org.jetbrains.kotlin.idea.debugger.test.ArtUtils.getTestTimeoutMillis
-import org.jetbrains.kotlin.idea.debugger.test.ArtUtils.runTestOnArt
 import org.jetbrains.kotlin.idea.debugger.test.preference.*
 import org.jetbrains.kotlin.idea.debugger.test.util.BreakpointCreator
 import org.jetbrains.kotlin.idea.debugger.test.util.KotlinOutputChecker
@@ -71,8 +63,6 @@ import org.jetbrains.kotlin.psi.psiUtil.containingClassOrObject
 import org.jetbrains.kotlin.test.TargetBackend
 import org.junit.ComparisonFailure
 import java.io.File
-import java.lang.ProcessBuilder.Redirect.PIPE
-import kotlin.io.path.pathString
 
 internal const val KOTLIN_LIBRARY_NAME = "KotlinJavaRuntime"
 internal const val TEST_LIBRARY_NAME = "TestLibrary"
@@ -105,6 +95,7 @@ abstract class KotlinDescriptorTestCase : DescriptorTestCase(),
     private var logPropagator: LogPropagator? = null
 
     private var oldValues: OldValuesStorage? = null
+    private val vmAttacher = VmAttacher.getInstance()
 
     override fun runBare(testRunnable: ThrowableRunnable<Throwable>) {
         testAppDirectory = tmpDir("debuggerTestSources")
@@ -181,6 +172,9 @@ abstract class KotlinDescriptorTestCase : DescriptorTestCase(),
         atDebuggerTearDown { invokeAndWaitIfNeeded { oldValues?.revertValues() } }
         atDebuggerTearDown { KotlinEvaluator.LOG_COMPILATIONS = false }
         atDebuggerTearDown { restoreIdeCompilerSettings() }
+
+        vmAttacher.setUp()
+        atDebuggerTearDown { vmAttacher.tearDown() }
     }
 
     protected fun dataFile(fileName: String): File = File(getTestDataPath(), fileName)
@@ -372,10 +366,8 @@ abstract class KotlinDescriptorTestCase : DescriptorTestCase(),
             .runProfile(MockConfiguration(myProject))
             .build()
 
-        val debuggerSession = when (runTestOnArt()) {
-            true -> createArtLocalProcess(javaParameters, environment)
-            false -> createJvmLocalProcess(javaParameters, environment)
-        }
+        environment.putUserData(DefaultDebugEnvironment.DEBUGGER_TRACE_MODE, traceMode)
+        val debuggerSession = vmAttacher.attachVirtualMachine(this, javaParameters, environment)
 
         val processHandler = debuggerSession.process.processHandler
         debuggerSession.process.addProcessListener(object : ProcessAdapter() {
@@ -406,68 +398,6 @@ abstract class KotlinDescriptorTestCase : DescriptorTestCase(),
         assertNotNull(process)
 
         return debuggerSession
-    }
-
-    private fun createJvmLocalProcess(javaParameters: JavaParameters, environment: ExecutionEnvironment): DebuggerSession {
-        val debuggerRunnerSettings = (environment.runnerSettings as GenericDebuggerRunnerSettings)
-        val javaCommandLineState: JavaCommandLineState = object : JavaCommandLineState(environment) {
-            override fun createJavaParameters() = javaParameters
-
-            override fun createTargetedCommandLine(request: TargetEnvironmentRequest): TargetedCommandLineBuilder {
-                return getJavaParameters().toCommandLine(request)
-            }
-        }
-
-        val debugParameters =
-            RemoteConnectionBuilder(
-                debuggerRunnerSettings.LOCAL,
-                debuggerRunnerSettings.transport,
-                debuggerRunnerSettings.debugPort
-            )
-                .checkValidity(true)
-                .asyncAgent(true)
-                .create(javaCommandLineState.javaParameters)
-
-        val env = javaCommandLineState.environment
-        env.putUserData(DefaultDebugEnvironment.DEBUGGER_TRACE_MODE, traceMode)
-
-        return attachVirtualMachine(javaCommandLineState, env, debugParameters, false)
-    }
-
-    private fun createArtLocalProcess(javaParameters: JavaParameters, environment: ExecutionEnvironment): DebuggerSession {
-        println("Running on ART VM")
-        setTimeout(getTestTimeoutMillis())
-        val mainClass = javaParameters.mainClass
-        val dexFile = buildDexFile(javaParameters.classPath.pathList)
-        val command = buildCommandLine(dexFile.pathString, mainClass)
-        testRootDisposable.whenDisposed {
-            dexFile.delete()
-        }
-        val art = ProcessBuilder()
-            .command(command)
-            .redirectOutput(PIPE)
-            .start()
-
-        val port: String = art.inputStream.bufferedReader().use {
-            while (true) {
-                val line = it.readLine() ?: break
-                if (line.startsWith("Listening for transport")) {
-                    val port = line.substringAfterLast(" ")
-                    return@use port
-                }
-            }
-            throw IllegalStateException("Failed to read listening port from ART")
-        }
-
-        val debugParameters =
-            RemoteConnectionBuilder(false, DebuggerSettings.SOCKET_TRANSPORT, port)
-                .checkValidity(true)
-                .asyncAgent(true)
-                .create(javaParameters)
-
-        val remoteState = RemoteStateState(project, debugParameters)
-
-        return attachVirtualMachine(remoteState, environment, debugParameters, false)
     }
 
     open fun addMavenDependency(compilerFacility: DebuggerTestCompilerFacility, library: String) {
